Got rid of ValidationContext.kt
Still unsure if this is a good approach. On the one hand I like how upfront this forces me to be about my validation, on the other hand it looks overly verbose and I feel like I'm missing a trick here to make it more readable.

I imagine this would be simpler to process when sealed class implementors are allowed to be in separate files but same package